Move ClusterInstance to mode-core and rename to Instance
