+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/test/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/registry/status/compute/service/ComputeNodeStatusServiceTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.mode.manager.cluster.coordinator.registry.status.compute.service;
 
-import org.apache.shardingsphere.mode.manager.cluster.coordinator.ClusterInstance;
+import org.apache.shardingsphere.mode.instance.Instance;
 import org.apache.shardingsphere.mode.repository.cluster.ClusterPersistRepository;
 import org.junit.Before;
 import org.junit.Test;
@@ -35,12 +35,12 @@ public final class ComputeNodeStatusServiceTest {
     
     @Before
     public void setUp() {
-        ClusterInstance.getInstance().init(3307);
+        Instance.getInstance().init(3307);
     }
     
     @Test
     public void assertRegisterOnline() {
         new ComputeNodeStatusService(repository).registerOnline();
-        verify(repository).persistEphemeral("/status/compute_nodes/online/" + ClusterInstance.getInstance().getId(), "");
+        verify(repository).persistEphemeral("/status/compute_nodes/online/" + Instance.getInstance().getId(), "");
     }
 }
